Outbox implemented for kafka messaging consistensy and Transactional also included (#3)
Implemented Outbox mechanism in order to ensure the consistency when new entities are saved or updated with Kafka messages. For instance: when a new follow request is saved and the Kafka producer fails to send the message, the event is in the table OutbooxEvent with its status PENDING. So it is scheduled that the method processOutboxEvents() runs every 5 seconds and sends the message.